+++ NEWS.md
@@ -90,6 +90,8 @@ absent or too weak.
 * When fetching JWKS, if `key_ops` is present on keys, only keys with `key_ops` 
 including `"verify"` are considered.
 
+* Token type is now validated before making a call to the userinfo endpoint.
+
 ## Fixed
 
 * Package now correctly requires `httr2` >= 1.1.0.